Handler

Documentation

getVarsR :: Handler TypedContent

Handle /variables GET requests by returning either all variables, or, if the name query parameter exists, the variable picked out by that parameter (if it exists).

getVarR :: Text -> Handler TypedContent

Handle /variables/{name} GET requests for a single variable name by returning the variable with that name, if it exists. If it does not exist, a 404 error response is returned.

postVarsR :: Handler ()

Handle /variables POST requests by creating or updating multiple variables at once. The variables to be updated are specified in the request body, which is a JSON object whose keys are variable names and values are the corresponding values. E.g.

{"a": 10, "b": -0.5}

putVarR :: Text -> Handler ()

Handle /variables/{name} PUT requests by creating or updating a single variable. The name of the variable to be updated is specified in the request URL slug, and the value to assign to that variable is represented in the request body as a JSON object of the form

{"value": 100}

deleteVarR :: Text -> Handler ()

Handle /variables/{name} DELETE requests by deleting the variable with the given name. If a variable with the given name doesn't exist, nothing happens.

postEvalR :: Handler TypedContent

Handle / POST requests by evaluating the expression in the request body, which should be a JSON object like

{"expr": "a*b + 100/(c-d)"}

If any of the variables occurring in the expression aren't defined in the variable store, then a 404 is returned.